Support for an exclusion filter to exclude files from tab transformation
the parameter tab_to_spaces allow to select file for which to
transform tab into spaces.
This is usually done based on the extension of the filename
there are case when specifics files that otherwise should be candidates
to a tab-to-space need to be excluded from such transforamtion
Specifically this was needed for .mk file. they use to be dmake files
but now we also have some that are gnu-make file, and therefore
the tab are semantically significants and muxt not be converted.

a new option
:set exclude_tabs=<regexp>
identify files to be excluded from tab conversion.
the rule is:
if a file has tab as leading characters in at least one line
and
the filename is matched by the tab_to_spaces pattern
and
(there is not exclusion or the filename is not match by the exclude regex)

then the conversion is applied.